refactor: create database package
