+++ app/src/main/java/dev/jefvda/smartplanner/DayOverviewFragment.kt
@@ -9,6 +9,8 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.navigation.fragment.navArgs
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import dev.jefvda.smartplanner.database.Activity
+import dev.jefvda.smartplanner.database.Weekday
 import dev.jefvda.smartplanner.databinding.FragmentDayOverviewBinding
 import java.util.*
