+++ app/src/main/java/dev/jefvda/smartplanner/WeekdayClickListener.kt
@@ -1,5 +1,7 @@
 package dev.jefvda.smartplanner
 
+import dev.jefvda.smartplanner.database.Weekday
+
 fun interface WeekdayClickListener {
     fun weekdayClicked(weekday: Weekday)
 }
