+++ app/src/main/java/dev/jefvda/smartplanner/WeekdayListAdapter.kt
@@ -3,6 +3,7 @@ package dev.jefvda.smartplanner
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import dev.jefvda.smartplanner.database.Weekday
 
 class WeekdayListAdapter(var weekdayList: MutableList<Weekday>, private val clickListener: WeekdayClickListener): RecyclerView.Adapter<WeekdayViewHolder>() {
